Remove special move behaviour.
This behaviour was perhaps slightly unexpected, and certainly out of
line with the standard library wrapper observer_ptr, which is almost
the same as neo::ptr. An added advantage of using the default special
functions is that Neo types are now trivially copyable (i.e. you can
copy them using std::memcpy).